backend/requirements_minimal.txt
fastapi==0.116.1
python-multipart==0.0.20
uvicorn==0.35.0
numpy==1.24.3
opencv-python==4.8.0.74
mediapipe==0.10.14
pillow==10.0.0
python-dotenv==1.0.0
rembg==2.0.50
pymongo==4.5.0
bcrypt==4.0.1
pyjwt==2.8.0
torch>=2.0.0
torchvision>=0.15.0
scikit-learn==1.3.0